requirements.txt
fastapi
uvicorn
groq
pypdf
python-docx
python-dotenv
requests
pydantic
